pyproject.toml
[project]
name = "openbiodiv-ichatbio-agent"
version = "0.1.0"
readme = "README.md"
description = "Agent to Agent (A2A) Protocol that serves as a bridge between OpenBiodiv's biodiversity data and iChatBio's agent ecosystem"
keywords = ["iChatBio", "OpenBioDiv", "biodiversity", "research", "science", "AI", "pensoft", "A2A"]
requires-python = ">=3.12,<3.14"

dependencies = [
    "ichatbio-sdk>=0.2.2",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[project.scripts]
openbiodiv-agent = "src.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src", "tests"]
log_cli = true
log_cli_level = "INFO"
asyncio_mode = "auto"
